Set password file permission to read only for user
Currently the password file permission is set to 0666 through the
call to os.Create(). This means that the user, group and others can
read and write this file which is a security issue as only the user
should have access to it. In addition to this it is only necessary to
allow the user to read the file only since write can lead to
accidental overwrites. Hence the permission of this file needs to be
set to 0400.