Change from canvas.Text to widget.Label
In the current implementation when the theme is changed the individual
tabs will be refreshed. Thus reflecting the change of the theme.
However, the canvas.Text does not update since the internal state
needs to be explicitly set and then refreshed. However, doing this
will be inefficient.

A better way is to use Labels. This will automatically update to the
appropriate font colour depending on the theme when the canvasObject of
the tab is refreshed, which already is implemented in setAppTheme().