Ensure files to be saved at the correct path
In the current implementation when the application is launched from
the same directory as the application binary the config and passwords
file were saving in the same directory as the application path.
However, when the application is launched from other places, for
example via a key binding, the config and password files do not get
saved in the same directory as the application binary path.

To fix this issue the current directory of the binary is found at run
time and the files will be read and written into that directory.