Fix service not being able to be selected
With the current implementation when a service is selected and the
user later uses the search field, the previously selected service is
still selected. This means that the new service at the same position
as the old service cannot be selected. Thus preventing in the correct
response from taking being. This is an issue that exists in the
search and delete tabs.