Fix information not getting cleared
When a service is already selected and the search field is used to
find another service the information displayed on the right in the
search and delete tabs do not get cleared/updated. Hence, to give
a better representation, the information on the right hand side
will be cleared when the user makes a new search.